requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0
mysql-connector-python>=9.0.0
requests>=2.32.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.0
lxml>=5.0.0
pandas>=2.0.0
akshare==1.17.99
curl_cffi==0.13.0
aiohttp>=3.9.0
tqdm>=4.60.0
openpyxl>=3.0.3
tabulate>=0.8.6
xlrd>=1.2.0
html5lib>=1.0.1
jsonpath>=0.82
nest_asyncio>=1.6.0
decorator>=4.4.2
mini-racer>=0.12.4